Adding DSL traits to aggregate implicits so users of scala-rules can extend the DSL more easily
