+++ engine/src/main/scala/org/scalarules/dsl/nl/grammar/package.scala
@@ -1,15 +1,5 @@
 package org.scalarules.dsl.nl
 
-import org.scalarules.dsl.nl.datum.DatumImplicits
-import org.scalarules.engine._
-
-package object grammar extends DslConditionImplicits
-    with DslEvaluationImplicits
-    with DatumImplicits {
-
-  type ConditionFunction = (Condition, Condition) => Condition
-
-  // Entrypoint for the DSL
-  def Gegeven(condition: DslCondition): GegevenWord = new GegevenWord(condition) //scalastyle:ignore method.name
+package object grammar extends ScalaRulesDsl {
 
 }
